Fix TestCalculateCrossPackageImportPath test expectation
The test expected "../common/types/v1/interfaces" for import from "common" to
"common.types.v1", but the correct relative path is "./types/v1/interfaces"
since we're importing from a child subdirectory.

The original expectation would only make sense if we needed to go up a level
and back down, which is not the case for parent-to-child relationships.

Fixed: Changed expectation to match actual behavior (./types/v1/interfaces)